Move base,js and polyfills.js to src/app/base
